Add Vector, Position & GameObject classes
Add Vector class, containing variables x and y, representing the x- and y-components of a 2D vector, and methods for adding another vector and multiplying by a scalar.
Add Position class, inheriting from the Vector class, and containing additional variables, properties and functions for checking the x- and y-values are within a certain range (e.g. x- and y-values both between 0 and 7 inclusive to ensure the vector corresponds to a valid LED in the 8x8 matrix).
Add GameObject class for representing objects in the game (e.g. the player).